requirements.txt
Flask==0.9
Flask-Bootstrap==2.2.2-1
Flask-Login==0.1.3
Flask-PyMongo==0.2.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.2
Jinja2==2.6
argparse==1.2.1
wsgiref==0.1.2
pillow
pymongo==2.5.1
WTForms==1.0.3
Werkzeug==0.8.3